+++ scripts/docs/configs/template-filters.js
@@ -1,0 +1,12 @@
+module.exports = function(templateEngine) {
+  // Nunjucks and Angular conflict in their template bindings so change the Nunjucks
+  // Also conflict with Jekyll
+  templateEngine.config.tags = {
+    variableStart: '<$',
+    variableEnd: '$>',
+    blockStart: '<@',
+    blockEnd: '@>',
+    commentStart: '<#',
+    commentEnd: '#>'
+  };
+};