+++ scripts/docs/filters/capital.js
@@ -1,7 +1,5 @@
+"use strict";
 module.exports = {
   name: 'capital',
-  process: function(str) {
-    str || (str = '');
-    return str.charAt(0).toUpperCase() + str.substring(1);
-  }
+  process: str => str? str.charAt(0).toUpperCase() + str.substring(1) : ''
 };