+++ scripts/docs/processors/npm-id.js
@@ -1,24 +1,19 @@
+"use strict";
 module.exports = function npmId(renderDocsProcessor) {
   return {
     name: 'npm-id',
     $runAfter: ['paths-computed'],
     $runBefore: ['rendering-docs'],
-    $process: function(docs) {
-      var currentVersion = renderDocsProcessor.extraData.version.current.name;
-
+    $process: docs => {
       // pretty up and sort the docs object for menu generation
       docs = docs.filter(function(doc) {
         return (!!doc.name && !!doc.outputPath) || doc.docType === 'index-page';
       });
 
-      docs.forEach(function(doc, i) {
-        doc.npmId = doc.outputPath.replace('/README.md', '') // for readmes
-          .replace('@ionic-native/plugins/','') // for readmes
-          .replace('content/docs/v2/native/', '') // for docs
-          .replace('/index.md', ''); // for docs
+      docs.forEach(doc => {
+        doc.npmId = doc.id.match(/plugins\/(.*)\/index/)[1];
       });
 
-      // returning docs will replace docs object in the next process
       return docs;
     }
   };