+++ src/@ionic-native/core/decorators.ts
@@ -203,15 +203,19 @@ export function Plugin(config: PluginConfig) {
     cls['getPluginName'] = function() {
       return config.pluginName;
     };
+
     cls['getPluginRef'] = function() {
       return config.pluginRef;
     };
+
     cls['getPluginInstallName'] = function() {
       return config.plugin;
     };
+
     cls['getPluginRepo'] = function() {
       return config.repo;
     };
+
     cls['getSupportedPlatforms'] = function() {
       return config.platforms;
     };
@@ -259,6 +263,7 @@ export function CordovaInstance(opts: any = {}) {
  */
 export function CordovaProperty(target: any, key: string) {
   Object.defineProperty(target, key, {
+    enumerable: true,
     get: () => {
       if (checkAvailability(target, key) === true) {
         return getPlugin(target.constructor.getPluginRef())[key];
@@ -282,6 +287,7 @@ export function CordovaProperty(target: any, key: string) {
  */
 export function InstanceProperty(target: any, key: string) {
   Object.defineProperty(target, key, {
+    enumerable: true,
     get: function(){
       return this._objectInstance[key];
     },