+++ src/@ionic-native/plugins/location-accuracy/index.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Plugin, Cordova } from '@ionic-native/core';
 
 /**
- * @name LocationAccuracy
+ * @name Location Accuracy
  * @description
  * This Cordova/Phonegap plugin for Android and iOS to request enabling/changing of Location Services by triggering a native dialog from within the app, avoiding the need for the user to leave your app to change location settings manually.
  *