+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+<a name="3.2.4"></a>
+## [3.2.4](https://github.com/driftyco/ionic-native/compare/v3.2.3...v3.2.4) (2017-03-26)
+
+
+
+<a name="3.2.3"></a>
+## [3.2.3](https://github.com/driftyco/ionic-native/compare/v3.2.2...v3.2.3) (2017-03-24)
+
+
+### Bug Fixes
+
+* **contacts:** fix crash if plugin is unavailable ([a9a511d](https://github.com/driftyco/ionic-native/commit/a9a511d)), closes [#1250](https://github.com/driftyco/ionic-native/issues/1250)
+* **core:** make properties enumerable ([a151295](https://github.com/driftyco/ionic-native/commit/a151295))
+
+
+
+<a name="3.2.2"></a>
+## [3.2.2](https://github.com/driftyco/ionic-native/compare/v3.2.1...v3.2.2) (2017-03-23)
+
+
+
 <a name="3.2.1"></a>
 ## [3.2.1](https://github.com/driftyco/ionic-native/compare/v3.2.0...v3.2.1) (2017-03-23)